# server/src/main/resources/openapi/storyblock-v1.yaml
openapi: 3.1.0
info:
  title: StoryBlock Engine API
  version: 1.1.0
  description: Immutable, evidence-bound narrative editing contracts.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /novels:
    post:
      operationId: createNovel
      summary: Create a novel from an initial canonical revision
      x-required-scopes: ["novel:admin"]
      parameters:
        - $ref: "#/components/parameters/IfMatchForCreation"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNovelRequest"
      responses:
        "201":
          description: Novel created
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NovelHead"
        default:
          $ref: "#/components/responses/Problem"
  /imports:
    post:
      operationId: importCanonicalPackage
      summary: Import a versioned canonical JSON document or package
      x-required-scopes: ["novel:admin"]
      parameters:
        - $ref: "#/components/parameters/IfMatchForCreation"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: Existing idempotent import result
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NovelHead"
        "201":
          description: Canonical package imported atomically
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NovelHead"
        default:
          $ref: "#/components/responses/Problem"
...
  /jobs/{jobId}:
    get:
      operationId: getJob
      summary: Get durable job status within its novel authorization boundary
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Durable job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        default:
          $ref: "#/components/responses/Problem"
  /artifacts/{artifactId}:
    get:
      operationId: downloadArtifact
      summary: Download a novel-scoped immutable result artifact
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/ArtifactId"
      responses:
        "200":
          description: Immutable artifact bytes
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            X-Artifact-Codec:
              $ref: "#/components/headers/ArtifactCodec"
            Content-Disposition:
              $ref: "#/components/headers/ContentDisposition"
          content:
            application/vnd.storyblock.package+json:
              schema:
                type: string
                format: binary
            application/vnd.storyblock.revision+json:
              schema:
                type: string
                format: binary
            application/vnd.storyblock.style-trace+json:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/Problem"
  /novels/{novelId}/exports:
    post:
      operationId: startExport
      summary: Submit a canonical JSON or package export job
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "202":
          $ref: "#/components/responses/JobAccepted"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: nv_<public-key-id>.<256-bit-secret>
  parameters:
    NovelId:
      name: novelId
      in: path
      required: true
      schema:
        type: string
        pattern: '^nov_[0-9a-f-]{36}$'
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        pattern: '^job_[0-9a-f-]{36}$'
...
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema:
        type: string
        pattern: '^art_[0-9a-f-]{36}$'
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: A single quoted strong canonical SHA-256 ETag.
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
    IfMatchForCreation:
      name: If-Match
      in: header
      required: true
      description: Wildcard precondition for creation in a collection.
      schema:
        const: '*'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
  headers:
    StrongEtag:
      description: Quoted strong ETag of the canonical revision or immutable resource view.
      required: true
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
    Location:
      description: Absolute or origin-relative durable resource URI.
      required: true
      schema:
        type: string
        format: uri-reference
    ArtifactCodec:
      description: Storage encoding applied to the returned bytes.
      required: true
      schema:
        type: string
        pattern: '^[a-z][a-z0-9.-]{1,63}$'
    ContentDisposition:
      description: Attachment filename for offline use.
      required: true
      schema:
        type: string
  responses:
    Problem:
      description: RFC problem response with a stable machine code
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    JobAccepted:
      description: Durable job accepted
      headers:
        Location:
          $ref: "#/components/headers/Location"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JobAccepted"
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,63}$'
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        current_revision_id:
          type: string
        current_etag:
          type: string
      additionalProperties: true
    CreateNovelRequest:
      type: object
      required: [initial_revision]
      properties:
        initial_revision:
          $ref: "#/components/schemas/CanonicalRevision"
      additionalProperties: false
    ImportRequest:
      type: object
      required: [format, document]
      properties:
        format:
          type: string
          enum: [canonical-revision, canonical-package]
        document:
          oneOf:
            - $ref: "#/components/schemas/CanonicalRevision"
            - $ref: "#/components/schemas/CanonicalPackage"
      additionalProperties: false
...
    NovelHead:
      type: object
      required: [novel_id, head_revision_id, head_sequence, head_hash, schema_version]
      properties:
        novel_id:
          type: string
        head_revision_id:
          type: string
        head_sequence:
          type: integer
          minimum: 0
        head_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        schema_version:
          type: string
      additionalProperties: false
    CanonicalRevision:
      type: object
      required: [schema_version, novel_id, revision_id, parent_revision_id, chapters, created_at, content_hash]
      properties:
        schema_version:
          const: 1.0.0
        novel_id:
          type: string
        revision_id:
          type: string
        parent_revision_id:
          type: [string, "null"]
        chapters:
          type: array
          items:
            type: object
            additionalProperties: true
        created_at:
          type: string
          format: date-time
        content_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        extensions:
          type: object
          additionalProperties: true
      additionalProperties: false
    CanonicalPackage:
      type: object
      required: [package_version, manifest, revisions, operations, artifacts]
      properties:
        package_version:
          const: 1.0.0
        manifest:
          $ref: "#/components/schemas/CanonicalPackageManifest"
        revisions:
          type: array
          minItems: 1
          maxItems: 100000
          items:
            $ref: "#/components/schemas/CanonicalPackageRevision"
        operations:
          type: array
          maxItems: 99999
          items:
            $ref: "#/components/schemas/CanonicalPackageOperation"
        artifacts:
          type: array
          maxItems: 10000
          items:
            $ref: "#/components/schemas/CanonicalPackageArtifact"
      additionalProperties: false
    CanonicalPackageManifest:
      type: object
      required: [novel_id, schema_version, head_revision_id, head_sequence, head_hash, revision_count, operation_count, artifact_count]
      properties:
        novel_id:
          type: string
        schema_version:
          const: 1.0.0
        head_revision_id:
          type: string
        head_sequence:
          type: integer
          minimum: 0
        head_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        revision_count:
          type: integer
          minimum: 1
        operation_count:
          type: integer
          minimum: 0
        artifact_count:
          type: integer
          minimum: 0
      additionalProperties: false
    CanonicalPackageRevision:
      type: object
      required: [sequence, document]
      properties:
        sequence:
          type: integer
          minimum: 0
        document:
          $ref: "#/components/schemas/CanonicalRevision"
      additionalProperties: false
    CanonicalPackageOperation:
      type: object
      required: [sequence, operation_hash, operation, result_revision_id, result_hash, committed_at]
      properties:
        sequence:
          type: integer
          minimum: 1
        operation_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        operation:
          $ref: "#/components/schemas/OperationEnvelope"
        result_revision_id:
          type: string
        result_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        committed_at:
          type: string
          format: date-time
      additionalProperties: false
    CanonicalPackageArtifact:
      type: object
      required: [artifact_id, revision_id, kind, media_type, codec, content_hash, size_bytes, created_at, content_base64]
      properties:
        artifact_id:
          type: string
        revision_id:
          type: string
        kind:
          type: string
        media_type:
          type: string
        codec:
          type: string
        content_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        size_bytes:
          type: integer
          minimum: 0
          maximum: 16777216
        created_at:
          type: string
          format: date-time
        content_base64:
          type: string
          contentEncoding: base64
      additionalProperties: false
...
    OperationEnvelope:
      type: object
      required: [operation_id, idempotency_key, novel_id, base_revision_id, expected_head_hash, type, payload]
      properties:
        operation_id:
          type: string
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 200
        novel_id:
          type: string
        base_revision_id:
          type: string
        expected_head_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        type:
          type: string
          enum: [insert_blocks, replace_block_range, delete_block_range, split_block, merge_blocks, extend_block, move_block_range, correct_block_meta, set_scene_initial_meta, restore_revision_content]
        payload:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    JobAccepted:
      type: object
      required: [job_id, status, status_uri]
      properties:
        job_id:
          type: string
        status:
          const: queued
        status_uri:
          type: string
          format: uri-reference
      additionalProperties: false
    Job:
      oneOf:
        - $ref: "#/components/schemas/ExportJob"
        - $ref: "#/components/schemas/StyleAnalysisJob"
    ExportJob:
      type: object
      required: [job_id, novel_id, revision_id, kind, status, attempt, result_artifact_id, result_uri, created_at, updated_at]
      properties:
        job_id:
          type: string
          pattern: '^job_[0-9a-f-]{36}$'
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        kind:
          const: canonical-export
        status:
          const: succeeded
        attempt:
          const: 1
        result_artifact_id:
          type: string
          pattern: '^art_[0-9a-f-]{36}$'
        result_uri:
          type: string
          format: uri-reference
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    StyleAnalysisJobFields:
      type: object
      required: [analysis_id, attempt, created_at, failure_code, job_id, kind, max_attempts, novel_id, result_artifact_id, result_hash, retention_until, revision_id, status, updated_at]
      properties:
        analysis_id:
          type: string
          pattern: '^ana_[0-9a-f-]{36}$'
        attempt:
          type: integer
          minimum: 0
          maximum: 20
        created_at:
          type: string
          format: date-time
        failure_code:
          type: [string, "null"]
          pattern: '^[a-z][a-z0-9._-]{1,63}$'
        job_id:
          type: string
          pattern: '^job_[0-9a-f-]{36}$'
        kind:
          const: style-analysis
        max_attempts:
          type: integer
          minimum: 1
          maximum: 20
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        result_artifact_id:
          type: [string, "null"]
          pattern: '^art_[0-9a-f-]{36}$'
        result_hash:
          type: [string, "null"]
          pattern: '^sha256:[0-9a-f]{64}$'
        retention_until:
          type: string
          format: date-time
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        updated_at:
          type: string
          format: date-time
    StyleAnalysisJob:
      allOf:
        - $ref: "#/components/schemas/StyleAnalysisJobFields"
      unevaluatedProperties: false
...
    ExportRequest:
      type: object
      required: [revision_id, format]
      properties:
        revision_id:
          type: string
        format:
          type: string
          enum: [canonical-revision, canonical-package]
      additionalProperties: false
...